requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0

python-dotenv==1.0.1

supabase==2.4.0

pinecone-client==3.2.2

sentence-transformers==2.7.0
torch==2.2.2
transformers==4.39.3

langchain==0.1.14
langchain-core==0.1.42
langchain-community==0.0.31
langchain-google-genai==0.0.12

ollama==0.1.8

pytesseract==0.3.10
Pillow==10.3.0
PyMuPDF==1.23.26

numpy==1.26.4

requests==2.31.0
